+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.1.0
+
+Added:
+
+- `NullableIterableExtensions.isEmpty` and `.isNotEmpty`
+- `NullableMapExensions.isEmpty` and `.isNotEmpty`
+
 # 1.0.0
 
 Breaking: